Move arg parsing to separate file
